+++ CHANGES.md
@@ -11,7 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.0.17
 
-- Use of `img/mathjslab-og.png` (1200x630) in Open Graph and Twitter records.
+- Use of `images/mathjslab-og.png` (1200x630) in Open Graph and Twitter records.
 - Introduction of the `seo.html` file, featuring links to online tools for technical SEO.
 
 ## 0.0.16